fix compiler warnings

+++ dngconvert/dngconvert.cpp
@@ -204,7 +204,7 @@ int main(int argc, const char* argv [])
     {
         negative->SetQuadMosaic(imgdata.idata.filters);
     }
-    else if (0 == memcmp("FUJIFILM", rawImage->MakeName().Get(), std::min((uint32)8, sizeof(rawImage->MakeName().Get()))))
+    else if (0 == memcmp("FUJIFILM", rawImage->MakeName().Get(), std::min((uint32)8, (uint32)sizeof(rawImage->MakeName().Get()))))
     {
         negative->SetFujiMosaic(0);
     }
@@ -229,10 +229,10 @@ int main(int argc, const char* argv [])
             negative->SetBayerMosaic(bayerPhase);
     }
 
-    negative->SetWhiteLevel(rawImage->WhiteLevel(0), 0);
-    negative->SetWhiteLevel(rawImage->WhiteLevel(1), 1);
-    negative->SetWhiteLevel(rawImage->WhiteLevel(2), 2);
-    negative->SetWhiteLevel(rawImage->WhiteLevel(3), 3);
+    negative->SetWhiteLevel((uint32)rawImage->WhiteLevel(0), 0);
+    negative->SetWhiteLevel((uint32)rawImage->WhiteLevel(1), 1);
+    negative->SetWhiteLevel((uint32)rawImage->WhiteLevel(2), 2);
+    negative->SetWhiteLevel((uint32)rawImage->WhiteLevel(3), 3);
 
     const dng_mosaic_info* mosaicinfo = negative->GetMosaicInfo();
     if ((mosaicinfo != NULL) && (mosaicinfo->fCFAPatternSize == dng_point(2, 2)))
@@ -395,9 +395,9 @@ int main(int argc, const char* argv [])
             streamPriv.Put(exiv2Meta.MakerNoteByteOrder().Get(), exiv2Meta.MakerNoteByteOrder().Length());
             streamPriv.Put_uint32(exiv2Meta.MakerNoteOffset());
             streamPriv.Put(exiv2Meta.MakerNoteData(), exiv2Meta.MakerNoteLength());
-            AutoPtr<dng_memory_block> blockPriv(host.Allocate(streamPriv.Length()));
+            AutoPtr<dng_memory_block> blockPriv(host.Allocate((uint32)streamPriv.Length()));
             streamPriv.SetReadPosition(0);
-            streamPriv.Get(blockPriv->Buffer(), streamPriv.Length());
+            streamPriv.Get(blockPriv->Buffer(), (uint32)streamPriv.Length());
             negative->SetPrivateData(blockPriv);
         }
     }
@@ -481,9 +481,9 @@ int main(int argc, const char* argv [])
         embedDataStream.Put_uint32(0);
         embedDataStream.Put_uint32(0);
 
-        AutoPtr<dng_memory_block> block(host.Allocate(embedDataStream.Length()));
+        AutoPtr<dng_memory_block> block(host.Allocate((uint32)embedDataStream.Length()));
         embedDataStream.SetReadPosition(0);
-        embedDataStream.Get(block->Buffer(), embedDataStream.Length());
+        embedDataStream.Get(block->Buffer(), (uint32)embedDataStream.Length());
 
         dng_md5_printer md5;
         md5.Process(block->Buffer(), block->LogicalSize());
@@ -527,8 +527,8 @@ int main(int argc, const char* argv [])
     jpeg_preview->fPhotometricInterpretation = piYCbCr;
     jpeg_preview->fPreviewSize               = jpegImage->Size();
     jpeg_preview->fYCbCrSubSampling          = dng_point(2, 2);
-    jpeg_preview->fCompressedData.Reset(host.Allocate(dms->Length()));
-    dms->Get(jpeg_preview->fCompressedData->Buffer_char(), dms->Length());
+    jpeg_preview->fCompressedData.Reset(host.Allocate((uint32)dms->Length()));
+    dms->Get(jpeg_preview->fCompressedData->Buffer_char(), (uint32)dms->Length());
     jpeg_preview->fInfo.fApplicationName.Set_ASCII("DNG SDK");
     jpeg_preview->fInfo.fApplicationVersion.Set_ASCII("1.3");
     //jpeg_preview->fInfo.fDateTime = ;
